odm/requirements.txt
torch>=1.10.0
torchvision>=0.11.1
numpy>=1.20.0
matplotlib>=3.4.0
pillow>=8.0.0
tqdm>=4.60.0
tensorboard>=2.5.0
seaborn>=0.11.0
pycocotools>=2.0.4
scikit-learn>=1.0.0
argparse>=1.1 